pyInterface: Load only needed amplitudes in calcEventWeights
Do not load amplitudes where the corresponding transition amplitude is
0.0, i.e. waves which were fixed to 0.0 during the fit.

Implement the calcEventWeights script to calculate weights for MC events
from the precalculated amplitudes based on a given fit result.
The weights are calculated for a given integral multibin.

Additionally, ignore pylint error (E1111(assignment-from-no-return) in
`calcEventWeights` as this gives an false-positive here.