pyInterface: Update of plotAngles
Implement the possibility to pass an integralMultibinID to the script.
Events of the input file will only be taken into account if they lie
within the intebralMultibin.
Additionally, the old loop over bins was removed as it was obsolete.
Additionally, the output file is not overwritten if it exists, but the
uses will be asked whether the file should be overwritten.