storageFormats: Increase limit for autosave when writing amplitude trees
Increase the limit for autoflush and autosave to 600MB when saving
amplitude trees to avoid that multiple keys are stored in the amplitude
root-file.
These multiple keys lead, for whatever reason, to a very large memory
consumption when reading the amplitude files for calculating the
integral matrices.
Additionally, reduce the read cache for the amplitude trees when
calculating the integral matrices to 100MB and scale the cache per tree
with the total number of waves.