decayAmplitude: Implement high-performance `addEvent` function for `ampIntegralMatrix`
Implement an `addEvent` member function to `ampIntegralMatrix`, which
takes a vector<complex<double>> as argument instead of a
map<string,complex<double>>. This function is much faster than the
other one as no checks are performed and the wave index is used instead
of the wave name. However, the user has to give the amplitudes in the
correct order, otherwise the matrix is screwed up.
This improves the performance of the on-the-fly integration by 17%.
(50000 events, 95 waves)