partialWaveFit: Improve performance if `intensity`, `spinDensityMatrixElem`
Improve the performance of the `intensity` by directly using the
amplitudes and not the spin-density matrix elements.
This reduced the matching of wave names and amplitude names, which is
very expensive, especially for spin-total intensity calculation.

Improve the performance of `spinDensityMatrixElem` for the diagonal
elements by building the amplitude-index vector only once, which is very
expensive.