pyInterface: Store empty amplitude tree for event files without events
It can happen that event files do not have events, because there are no
events in the corresponding bin.
So far, `calcAmplitudes` did neither store an empty amplitude tree nor
an `amplitudeMetadata` object in the corresponding amplitude file.
Now, an empty amplitude tree and the corresponding `amplitudeMetadata`
object is stored.