decayAmplitude: Fix memory leak in `ampIntegralMatrix`
The matrix is stored as an array in the ROOT file.
If the matrix is loaded from a file, the data from the array is copied
into the boost matrix.
However, after the copy, the array is not deleted, which leads to a
permanent doubling of the memory consumption.
Now, the memory consumption is only doubled during the load.